Issue 95: Validation on numeric textfields stops long numbers being allowed

+++ trunk/microemulator/microemu-midp/src/main/java/org/microemu/device/InputMethod.java
@@ -107,7 +107,7 @@ public abstract class InputMethod
             case TextField.NUMERIC :
                 if (text != null && text.length() > 0 && !text.equals("-")) {
                     try { 
-                        Integer.parseInt(text); 
+                        Long.parseLong(text); 
                     } catch (NumberFormatException e) { 
                         return false;
                     }
